Switched back to EmailJS but with Mailgun
So after some issues with Gmail + EmailJS, SMTP2GO + EmailJS, Mailjet + EmailJS, and Mailjet setups, I finally got it working with Mailgun + EmailJS.
One of the issues was that an email address was present in the From Name, which caused many issues with Mailgun, but it did not have the limitations other providers had.

This process will be further documented in the docs.